do_x

+++ test.c
@@ -4,8 +4,8 @@ int main()
 {
 	char *c = "aaaaaa";
 	char **str;
-	size_t d = 111;
+	int d = 111;
 
-	ft_printf("%.20u 2\n", d);
+	ft_printf("%.20d 2\n", d);
 	return 0;
 }
